+++ test/dummy/config/routes.rb
@@ -47,6 +47,7 @@ Rails.application.routes.draw do
   delete "shopify_plugin_demo/connect", to: "shopify_plugin_demo/connections#destroy", as: :shopify_plugin_demo_disconnect
   post "shopify_plugin_demo/uninstall", to: "shopify_plugin_demo/uninstalls#create"
   get "shopify_plugin_demo/storefront/embed", to: "shopify_plugin_demo/storefront_embeds#show"
+  get "shopify_plugin_demo/storefront/embed_boot.js", to: "shopify_plugin_demo/storefront_embeds#boot"
 
   root "home#index"